Fixed infinite run issue
The simulator would get stuck running forever after the 16th or so iteration.
After profiling the code, it seems like the code was getting stuck in `subs`

To fix this, I replaced all subs with matlab functions that take
in variables that were meant to be substituted as input.

This seems to make the script run faster and prevent the code from
getting stuck.